requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.40.0
torch==2.1.2
PyMuPDF==1.21.1
